requirements.txt
discord.py
oauthlib
pyshorteners
python-dateutil
pytz
requests
requests-oauthlib
